Separate creation of courses into separate cog
