+++ src/main.py
@@ -11,7 +11,8 @@ EXTENSIONS = [
     "aoun",
     "april_fools",
     "clear",
-    "course_registration",
+    "course_regisration.course_registration",
+    "course_regisration.create_course",
     "day",
     "help",
     "hours.hours",
